Fixed memchr and putnbr. WIP on strncat.
putnbr uses a cast to long int to allow for negative integers to be
handled correctly. Memchr problem seems to have been the linux
environment in testing, and not really the function itself.
Strncat has a problem with it's null termination, as strncpy
has the possibility of not including the terminating byte.